RDKMVE-3147: fbttestnative - fix the runtime and build error

+++ firebolt-test-app/native/src/utils.h
@@ -183,7 +183,7 @@ public:
 
     const std::string&              name()    const { return name_; }
     const std::vector<std::string>& methods() const { return methods_; }
-    const std::int& methodCount() const { return static_cast<std::int>(methods_.size()); }
+    const std::size_t& methodCount() const { return static_cast<std::size_t>(methods_.size()); }
 
     virtual void runMethod(const std::string& method) = 0;
